yarn install axios, react-redux, node-sass. rename file.
